Pharmasoft/server/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-MySQLdb==0.2.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
PyYAML==5.4.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
